Add no appointment found test.
